requirements.txt
fastapi
uvicorn
faiss-cpu
openai
numpy
python-multipart